SP-2 T3: extract model_turn translator + rename SpanID/Ts/Surface fields
